Fix for Windows absolute drive paths
Currently, when used with absolute drive paths, this results in a weird tree of paths that begin with "c:" but no backslash before the first directory.  This is due to the weird way that node.js handles `path.normalize("c:")`, returning `"c:."`.

This resolves the issue by ensuring that if an absolute drive path is given, it uses `"c:\\"` instead.